crawler:: 최적화/정제 과정 분리
